New generated code updates

+++ application/frontend-react/edemokracia__admin__admin/src/pages/admin/admin/debates/view/actions/pageDeleteDebates.tsx
@@ -4,30 +4,44 @@
 // Factory expression: #getActionsForPages(#application)
 // Path expression: #pagePath(#self.value)+'actions/'+#pageActionPathSuffix(#self.key,#self.value)+'.tsx'
 // Template name: actor/src/pages/actions/action.tsx
-// Base URL: mvn:hu.blackbelt.judo.generator:judo-ui-react:1.0.0.20230413_041932_3a0d360a_develop
+// Base URL: mvn:hu.blackbelt.judo.generator:judo-ui-react:1.0.0.20230413_174054_1b98627b_develop
 // Template file: actor/src/pages/actions/action.tsx.hbs
 // Action: DeleteAction
 
+import { useTranslation } from 'react-i18next';
 import type { JudoIdentifiable } from '@judo/data-api-common';
 import { OBJECTCLASS } from '@pandino/pandino-api';
 import { useSnackbar } from 'notistack';
 import { useErrorHandler, ERROR_PROCESSOR_HOOK_INTERFACE_KEY } from '../../../../../../utilities';
+import { useConfirmDialog } from '../../../../../../components/dialog';
 import { AdminDebate, AdminDebateStored, AdminDebateQueryCustomizer } from '../../../../../../generated/data-api';
 import { adminAdminServiceForDebatesImpl, adminDebateServiceImpl } from '../../../../../../generated/data-axios';
 
 export type PageDeleteDebatesAction = () => (selected: AdminDebateStored, successCallback: () => void) => Promise<void>;
 
 export const usePageDeleteDebatesAction: PageDeleteDebatesAction = () => {
+  const { t } = useTranslation();
   const { enqueueSnackbar } = useSnackbar();
+  const { openConfirmDialog } = useConfirmDialog();
   const handleActionError = useErrorHandler(
     `(&(${OBJECTCLASS}=${ERROR_PROCESSOR_HOOK_INTERFACE_KEY})(operation=RowDeleteAction))`,
   );
 
   return async function pageDeleteDebatesAction(selected: AdminDebateStored, successCallback: () => void) {
     try {
-      await adminDebateServiceImpl.delete(selected);
+      const confirmed = await openConfirmDialog(
+        'row-delete-action',
+        t('judo.modal.confirm.confirm-delete', {
+          defaultValue: 'Are you sure you would like to delete the selected element?',
+        }),
+        t('judo.modal.confirm.confirm-title', { defaultValue: 'Confirm action' }),
+      );
 
-      successCallback();
+      if (confirmed) {
+        await adminDebateServiceImpl.delete(selected);
+
+        successCallback();
+      }
     } catch (error) {
       handleActionError(error);
     }
